SMS Spam problem solved with Random Forest
Using the SMS Spam dataset from https://www.kaggle.com/mrlucasfischer/bertembedded-spam-messages , the spam classification problem was solved in R program with Random forest. The accuracy of the solution was shown with the confusion matrix, and the predictions for the ROC curve were output. Finally, the important variables of the solution are shown.